fix long

+++ Clansty.tianlang/Exceptions/DuplicateNameException.cs
@@ -8,7 +8,7 @@ namespace Clansty.tianlang
     class DuplicateNameException : Exception
     {
         public DuplicateNameException(): base(message: "姓名重名") { }
-        public DuplicateNameException(string name, int enrollment) : 
+        public DuplicateNameException(string name, long enrollment) : 
             base(message: $"姓名重名 姓名: {name}, 年级: {enrollment}") { }
         public DuplicateNameException(string name) : 
             base(message: $"姓名重名 姓名: {name}") { }
